Added Qix tests. Arranged tests. Renamed test file. Added "ext-readline": "*" to composer.json Minor index.php changes.
