+++ composer.json
@@ -4,5 +4,8 @@
   },
   "autoload": {
     "psr-4": {"App\\": "app/"}
+  },
+  "require": {
+    "ext-readline": "*"
   }
 }
